refactor(wp): move deprecated code to own file for clarity
